+++ data/duke.txt
@@ -1,1 +1,4 @@
-D | 0 | wahattt | 02 Dec 12:00
+T | 0 | hi
+T | 0 | hi
+D | 0 | hiii | 2020-02-01 12:00
+E | 0 | hiiii | 2020-01-01 00:00 | 2021-01-01 00:00
