+++ src/main/java/Deadline.java
@@ -4,14 +4,15 @@ import java.time.format.DateTimeFormatter;
 
 class Deadline extends Task {
     private LocalDateTime byDate;
-    public Deadline(String description, LocalDateTime byDate) {
+    public Deadline(String description, LocalDateTime byDate, boolean isMarked) {
         super(description);
         this.byDate = byDate;
+        super.isMarked = isMarked;
     }
 
-    public String changeFormat() {
+    public String writeFormat() {
         int isDone = isMarked ? 1 : 0;
-        String formattedDate = byDate.format(DateTimeFormatter.ofPattern("dd MMM HH:mm"));
+        String formattedDate = byDate.format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm"));
         return "D" + " | " + isDone + " | " + super.description + " | " + formattedDate;
     }
